+++ core/src/main/java/greencity/controller/ManagementOrderController.java
@@ -11,6 +11,7 @@ import greencity.filters.CertificatePage;
 import greencity.filters.OrderPage;
 import greencity.filters.OrderSearchCriteria;
 import greencity.service.ubs.CertificateService;
+import greencity.service.ubs.CoordinateService;
 import greencity.service.ubs.UBSManagementService;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
@@ -38,14 +39,17 @@ import java.util.Set;
 public class ManagementOrderController {
     private final UBSManagementService ubsManagementService;
     private final CertificateService certificateService;
+    private final CoordinateService coordinateService;
 
     /**
      * Constructor with parameters.
      */
     @Autowired
-    public ManagementOrderController(UBSManagementService ubsManagementService, CertificateService certificateService) {
+    public ManagementOrderController(UBSManagementService ubsManagementService, CertificateService certificateService,
+        CoordinateService coordinateService) {
         this.ubsManagementService = ubsManagementService;
         this.certificateService = certificateService;
+        this.coordinateService = coordinateService;
     }
 
     /**
@@ -106,7 +110,7 @@ public class ManagementOrderController {
     })
     @GetMapping("/all-undelivered")
     public ResponseEntity<List<GroupedOrderDto>> allUndeliveredCoords() {
-        return ResponseEntity.status(HttpStatus.OK).body(ubsManagementService.getAllUndeliveredOrdersWithLiters());
+        return ResponseEntity.status(HttpStatus.OK).body(coordinateService.getAllUndeliveredOrdersWithLiters());
     }
 
     /**
@@ -127,7 +131,7 @@ public class ManagementOrderController {
     public ResponseEntity<List<GroupedOrderDto>> groupCoords(@RequestParam Double radius,
         @RequestParam(required = false, defaultValue = "3000") Integer litres) {
         return ResponseEntity.status(HttpStatus.OK)
-            .body(ubsManagementService.getClusteredCoords(radius, litres));
+            .body(coordinateService.getClusteredCoords(radius, litres));
     }
 
     /**
@@ -147,7 +151,7 @@ public class ManagementOrderController {
         @RequestParam(required = false, defaultValue = "3000") Integer litres,
         @RequestParam(required = false, defaultValue = "0") Double additionalDistance) {
         return ResponseEntity.status(HttpStatus.OK)
-            .body(ubsManagementService.getClusteredCoordsAlongWithSpecified(specified, litres, additionalDistance));
+            .body(coordinateService.getClusteredCoordsAlongWithSpecified(specified, litres, additionalDistance));
     }
 
     /**
