+++ core/src/test/java/greencity/controller/ManagementOrderControllerTest.java
@@ -7,6 +7,7 @@ import greencity.dto.*;
 import greencity.filters.CertificateFilterCriteria;
 import greencity.filters.CertificatePage;
 import greencity.service.ubs.CertificateService;
+import greencity.service.ubs.CoordinateService;
 import greencity.service.ubs.UBSManagementService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -25,7 +26,6 @@ import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.validation.Validator;
 
 import java.security.Principal;
-import java.util.List;
 import java.util.Optional;
 
 import static greencity.ModelUtils.*;
@@ -43,6 +43,9 @@ class ManagementOrderControllerTest {
 
     @Mock
     UBSManagementService ubsManagementService;
+
+    @Mock
+    CoordinateService coordinateService;
 
     @Mock
     CertificateService certificateService;
@@ -358,7 +361,7 @@ class ManagementOrderControllerTest {
         mockMvc.perform(get(ubsLink + "/all-undelivered"))
             .andExpect(status().isOk());
 
-        verify(ubsManagementService).getAllUndeliveredOrdersWithLiters();
+        verify(coordinateService).getAllUndeliveredOrdersWithLiters();
     }
 
     @Test
