+++ service-api/src/main/java/greencity/service/ubs/UBSManagementService.java
@@ -9,40 +9,8 @@ import org.springframework.web.multipart.MultipartFile;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
 
 public interface UBSManagementService {
-    /**
-     * Method to group orders into clusters including summary litres and specified
-     * coordinates.
-     *
-     * @param specified          - list of {@link CoordinatesDto}.
-     * @param litres             - preferred amount of litres.
-     * @param additionalDistance - additional km to radius.
-     * @return List of {@link GroupedOrderDto} lists.
-     * @author Oleh Bilonizhka
-     */
-    List<GroupedOrderDto> getClusteredCoordsAlongWithSpecified(Set<CoordinatesDto> specified,
-        int litres, double additionalDistance);
-
-    /**
-     * Method to group orders into clusters including summary litres.
-     *
-     * @param distance - preferred distance for clusterization.
-     * @param litres   - preferred amount of litres.
-     * @return List of {@link GroupedOrderDto} lists.
-     * @author Oleh Bilonizhka
-     */
-    List<GroupedOrderDto> getClusteredCoords(double distance, int litres);
-
-    /**
-     * Method returns all undelivered orders including litres.
-     *
-     * @return List of {@link GroupedOrderDto} lists.
-     * @author Oleh Bilonizhka
-     */
-    List<GroupedOrderDto> getAllUndeliveredOrdersWithLiters();
-
     /**
      * Method returns payment info.
      *
