+++ service/src/main/java/greencity/service/ubs/CoordinateServiceImpl.java
@@ -1,0 +1,315 @@
+package greencity.service.ubs;
+
+import greencity.dto.CoordinatesDto;
+import greencity.dto.GroupedOrderDto;
+import greencity.dto.OrderDto;
+import greencity.entity.coords.Coordinates;
+import greencity.entity.order.Order;
+import greencity.exceptions.ActiveOrdersNotFoundException;
+import greencity.exceptions.IncorrectValueException;
+import greencity.repository.AddressRepository;
+import greencity.repository.OrderRepository;
+import lombok.RequiredArgsConstructor;
+import org.modelmapper.ModelMapper;
+
+import java.util.*;
+import java.util.stream.Collectors;
+
+import static greencity.constant.ErrorMessage.*;
+
+@org.springframework.stereotype.Service
+@RequiredArgsConstructor
+public class CoordinateServiceImpl implements CoordinateService {
+    private final AddressRepository addressRepository;
+    private final OrderRepository orderRepository;
+    private final ModelMapper modelMapper;
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public List<GroupedOrderDto> getAllUndeliveredOrdersWithLiters() {
+        Set<Coordinates> allCoords = addressRepository.undeliveredOrdersCoords();
+        List<Order> allOrders = getAllUndeliveredOrders();
+        List<GroupedOrderDto> allOrdersWithLitres = new ArrayList<>();
+        for (Coordinates temp : allCoords) {
+            int currentCoordinatesCapacity =
+                addressRepository.capacity(temp.getLatitude(), temp.getLongitude());
+            List<Order> currentCoordinatesOrders = allOrders.stream().filter(
+                o -> o.getUbsUser().getAddress().getCoordinates().equals(temp)).collect(Collectors.toList());
+            List<OrderDto> currentCoordinatesOrdersDto = currentCoordinatesOrders.stream()
+                .map(o -> modelMapper.map(o, OrderDto.class)).collect(Collectors.toList());
+            allOrdersWithLitres.add(GroupedOrderDto.builder()
+                .amountOfLitres(currentCoordinatesCapacity)
+                .groupOfOrders(currentCoordinatesOrdersDto)
+                .build());
+        }
+        return allOrdersWithLitres;
+    }
+
+    /**
+     * Method finds undelivered orders.
+     *
+     * @return List of {@link Order}
+     */
+    private List<Order> getAllUndeliveredOrders() {
+        List<Order> allCoords = orderRepository.undeliveredAddresses();
+        if (allCoords.isEmpty()) {
+            throw new ActiveOrdersNotFoundException(UNDELIVERED_ORDERS_NOT_FOUND);
+        }
+        return allCoords;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public List<GroupedOrderDto> getClusteredCoords(double distance, int litres) {
+        checkIfSpecifiedLitresAndDistancesAreValid(distance, litres);
+        Set<Coordinates> allCoords = addressRepository.undeliveredOrdersCoordsWithCapacityLimit(litres);
+        List<GroupedOrderDto> allClusters = new ArrayList<>();
+
+        while (!allCoords.isEmpty()) {
+            Optional<Coordinates> any = allCoords.stream().findAny();
+            mainBlockOfGetClusteredCoords(allCoords, distance, litres, any, allClusters);
+        }
+        return allClusters;
+    }
+
+    /**
+     * Method checks if entered parameters are valid.
+     *
+     * @param distance - preferred amount of litres.
+     * @param litres   - preferred search radius.
+     */
+    private void checkIfSpecifiedLitresAndDistancesAreValid(double distance, int litres) {
+        if (distance < 0 || distance > 20) {
+            throw new IncorrectValueException(INAVALID_DISTANCE_AMOUNT);
+        }
+        if (litres < 0 || litres > 10000) {
+            throw new IncorrectValueException(INAVALID_LITRES_AMOUNT);
+        }
+    }
+
+    private void mainBlockOfGetClusteredCoords(Set<Coordinates> allCoords, double distance,
+        int litres, Optional<Coordinates> any, List<GroupedOrderDto> allClusters) {
+        any.ifPresent(coordinates -> {
+            Coordinates currentlyCoord = coordinates;
+
+            Set<Coordinates> closeRelatives = getCoordinateCloseRelatives(distance,
+                allCoords, currentlyCoord);
+            Coordinates centralCoord = getNewCentralCoordinate(closeRelatives);
+
+            while (!centralCoord.equals(currentlyCoord)) {
+                currentlyCoord = centralCoord;
+                closeRelatives = getCoordinateCloseRelatives(distance, allCoords, currentlyCoord);
+                centralCoord = getNewCentralCoordinate(closeRelatives);
+            }
+            int amountOfLitresInCluster = 0;
+            for (Coordinates current : closeRelatives) {
+                int currentCoordinatesCapacity =
+                    addressRepository.capacity(current.getLatitude(), current.getLongitude());
+                amountOfLitresInCluster += currentCoordinatesCapacity;
+            }
+            if (amountOfLitresInCluster > litres) {
+                List<Coordinates> closeRelativesSorted = new ArrayList<>(closeRelatives);
+                closeRelativesSorted.sort(getComparatorByDistanceFromCenter(centralCoord));
+                int indexOfCoordToBeDeleted = -1;
+
+                while (amountOfLitresInCluster > litres) {
+                    Coordinates coordToBeDeleted = closeRelativesSorted.get(++indexOfCoordToBeDeleted);
+                    int anountOfLitresInCurrentOrder = addressRepository
+                        .capacity(coordToBeDeleted.getLatitude(), coordToBeDeleted.getLongitude());
+                    amountOfLitresInCluster -= anountOfLitresInCurrentOrder;
+                    closeRelatives.remove(coordToBeDeleted);
+                }
+            }
+            for (Coordinates grouped : closeRelatives) {
+                allCoords.remove(grouped);
+            }
+
+            // mapping coordinates to orderDto
+            getUndeliveredOrdersByGroupedCoordinates(closeRelatives,
+                amountOfLitresInCluster, allClusters);
+        });
+    }
+
+    /**
+     * Method defines and returns all coordinates in certain radius.
+     *
+     * @param distance       - preferred distance for clusterization.
+     * @param allCoords      - list of {@link Coordinates} which shows all
+     *                       unclustered coordinates.
+     * @param currentlyCoord - {@link Coordinates} - chosen start coordinates.
+     * @return list of {@link Coordinates} - start coordinates with it's
+     *         distant @relatives.
+     * @author Oleh Bilonizhka
+     */
+    private Set<Coordinates> getCoordinateCloseRelatives(double distance,
+        Set<Coordinates> allCoords, Coordinates currentlyCoord) {
+        Set<Coordinates> coordinateWithCloseRelativesList = new HashSet<>();
+
+        for (Coordinates checked : allCoords) {
+            if (distanceBetweenEarthCoordinates(currentlyCoord.getLatitude(), currentlyCoord.getLongitude(),
+                checked.getLatitude(), checked.getLongitude()) <= distance) {
+                coordinateWithCloseRelativesList.add(checked);
+            }
+        }
+
+        return coordinateWithCloseRelativesList;
+    }
+
+    /**
+     * Method to determine distance between 2 earth coordinates.
+     *
+     * @param lat1 {@link Double} - latitude of 1 coordinate.
+     * @param lon1 {@link Double} - longitude of 1 coordinate.
+     * @param lat2 {@link Double} - latitude of 2 coordinate.
+     * @param lon2 {@link Double} - longitude of 2 coordinate.
+     * @return {@link Integer} distance in meters.
+     */
+    private double distanceBetweenEarthCoordinates(double lat1, double lon1, double lat2, double lon2) {
+        double earthRadiusKm = 6371;
+
+        double radiansLatitude = degreesToRadians(lat2 - lat1);
+        double radiansLongitude = degreesToRadians(lon2 - lon1);
+
+        lat1 = degreesToRadians(lat1);
+        lat2 = degreesToRadians(lat2);
+
+        double a = Math.sin(radiansLatitude / 2) * Math.sin(radiansLatitude / 2)
+            + Math.sin(radiansLongitude / 2) * Math.sin(radiansLongitude / 2) * Math.cos(lat1) * Math.cos(lat2);
+        double c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
+        return earthRadiusKm * c;
+    }
+
+    /**
+     * Method to convert degrees to radians.
+     *
+     * @param degrees {@link Double} degrees.
+     * @return {@link Double} radians.
+     */
+    private double degreesToRadians(double degrees) {
+        return degrees * Math.PI / 180;
+    }
+
+    /**
+     * Method defines new central coordinate for existing ones.
+     *
+     * @param coordinateWithCloseRelatives list of {@link Coordinates}.
+     * @return {@link Coordinates} new central coordinate.
+     * @author Oleh Bilonizhka
+     */
+    private Coordinates getNewCentralCoordinate(Set<Coordinates> coordinateWithCloseRelatives) {
+        double sumLat = 0;
+        double sumLon = 0;
+        int amountOfCoords = coordinateWithCloseRelatives.size();
+
+        for (Coordinates checked : coordinateWithCloseRelatives) {
+            sumLat += checked.getLatitude();
+            sumLon += checked.getLongitude();
+        }
+
+        return new Coordinates(sumLat / amountOfCoords, sumLon / amountOfCoords);
+    }
+
+    /**
+     * Method returns coordinates comparator by theirs distance from center of
+     * cluster.
+     *
+     * @param centralCoord {@link Integer}.
+     * @return {@link Comparator} of Coordinates.
+     * @author Oleh Bilonizhka
+     */
+    private Comparator<Coordinates> getComparatorByDistanceFromCenter(Coordinates centralCoord) {
+        return (o1, o2) -> {
+            Double o1Int = distanceBetweenEarthCoordinates(o1.getLatitude(), o1.getLongitude(),
+                centralCoord.getLatitude(), centralCoord.getLongitude()) * 1000;
+
+            Double o2Int = distanceBetweenEarthCoordinates(o2.getLatitude(), o2.getLongitude(),
+                centralCoord.getLatitude(), centralCoord.getLongitude()) * 1000;
+
+            return o2Int.compareTo(o1Int);
+        };
+    }
+
+    private void getUndeliveredOrdersByGroupedCoordinates(Set<Coordinates> closeRelatives, int amountOfLitresInCluster,
+        List<GroupedOrderDto> allClusters) {
+        List<Order> orderslist = new ArrayList<>();
+        for (Coordinates coordinates : closeRelatives) {
+            List<Order> orders =
+                orderRepository.undeliveredOrdersGroupThem(coordinates.getLatitude(), coordinates.getLongitude());
+            orderslist.addAll(orders);
+        }
+        GroupedOrderDto cluster = new GroupedOrderDto();
+        cluster.setGroupOfOrders(
+            orderslist.stream().map(order -> modelMapper.map(order, OrderDto.class)).collect(Collectors.toList()));
+        cluster.setAmountOfLitres(amountOfLitresInCluster);
+        allClusters.add(cluster);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public List<GroupedOrderDto> getClusteredCoordsAlongWithSpecified(Set<CoordinatesDto> specified,
+        int litres, double additionalDistance) {
+        checkIfSpecifiedLitresAndDistancesAreValid(additionalDistance, litres);
+
+        Set<Coordinates> allCoords = addressRepository.undeliveredOrdersCoords();
+        Set<Coordinates> result = specified.stream()
+            .map(c -> modelMapper.map(c, Coordinates.class)).collect(Collectors.toSet());
+        for (Coordinates temp : result) {
+            if (!allCoords.contains(temp)) {
+                throw new IncorrectValueException(NO_SUCH_COORDINATES + temp.getLatitude()
+                    + ", " + temp.getLongitude());
+            }
+        }
+
+        Coordinates centralCoord = getNewCentralCoordinate(result);
+        int specifiedCoordsCapacity = 0;
+        double newRadius = 0;
+        for (Coordinates temp : result) {
+            double distanceFromCentralCoord = distanceBetweenEarthCoordinates(temp.getLatitude(), temp.getLongitude(),
+                centralCoord.getLatitude(), centralCoord.getLongitude());
+            if (distanceFromCentralCoord > newRadius) {
+                newRadius = distanceFromCentralCoord;
+            }
+            specifiedCoordsCapacity += addressRepository.capacity(temp.getLatitude(), temp.getLongitude());
+        }
+        newRadius += additionalDistance;
+
+        List<Coordinates> coordinatesInsideRadiusWithoutSpecifiedCoords = new ArrayList<>();
+        for (Coordinates temp : allCoords) {
+            double distanceFromCentralCoord = distanceBetweenEarthCoordinates(temp.getLatitude(), temp.getLongitude(),
+                centralCoord.getLatitude(), centralCoord.getLongitude());
+            if (distanceFromCentralCoord < newRadius) {
+                coordinatesInsideRadiusWithoutSpecifiedCoords.add(temp);
+            }
+        }
+        coordinatesInsideRadiusWithoutSpecifiedCoords.removeAll(result);
+
+        coordinatesInsideRadiusWithoutSpecifiedCoords.sort(getComparatorByDistanceFromCenter(centralCoord));
+        int amountOfLitresToFill = litres - specifiedCoordsCapacity;
+        double fill = 0;
+        int allCoordsCapacity = specifiedCoordsCapacity;
+        for (int i = coordinatesInsideRadiusWithoutSpecifiedCoords.size() - 1; i > -1; i--) {
+            Coordinates temp = coordinatesInsideRadiusWithoutSpecifiedCoords.get(i);
+            int capacity = addressRepository.capacity(temp.getLatitude(), temp.getLongitude());
+            if (fill < amountOfLitresToFill) {
+                if ((fill + capacity) <= amountOfLitresToFill) {
+                    fill += capacity;
+                    allCoordsCapacity += capacity;
+                    result.add(temp);
+                }
+            } else {
+                break;
+            }
+        }
+        List<GroupedOrderDto> groupedOrderDtos = new ArrayList<>();
+        getUndeliveredOrdersByGroupedCoordinates(result,
+            allCoordsCapacity, groupedOrderDtos);
+
+        return groupedOrderDtos;
+    }
+}
