+++ service/src/test/java/greencity/service/UBSManagementServiceImplTest.java
@@ -6,7 +6,6 @@ import greencity.client.RestClient;
 import greencity.constant.AppConstant;
 import greencity.constant.OrderHistory;
 import greencity.dto.*;
-import greencity.entity.coords.Coordinates;
 import greencity.entity.enums.OrderStatus;
 import greencity.entity.enums.SortingOrder;
 import greencity.entity.language.Language;
@@ -20,25 +19,18 @@ import greencity.entity.user.employee.Position;
 import greencity.entity.user.employee.ReceivingStation;
 import greencity.entity.user.ubs.Address;
 import greencity.exceptions.*;
-import greencity.filters.CertificateFilterCriteria;
-import greencity.filters.CertificatePage;
 import greencity.filters.OrderPage;
 import greencity.filters.OrderSearchCriteria;
 import greencity.repository.*;
 import greencity.service.ubs.*;
-import org.hibernate.mapping.Any;
-import org.junit.Assert;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import org.junit.jupiter.params.provider.EnumSource;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.invocation.InvocationOnMock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.mockito.stubbing.Answer;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeToken;
 import org.springframework.data.domain.*;
@@ -47,7 +39,6 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.time.LocalDateTime;
 import java.util.*;
-import java.util.stream.Collectors;
 
 import static greencity.ModelUtils.*;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -58,8 +49,6 @@ import static org.mockito.Mockito.*;
 class UBSManagementServiceImplTest {
     @Mock(lenient = true)
     AddressRepository addressRepository;
-    double distance = 2;
-    int litres = 1000;
 
     @Mock
     private FileService fileService;
@@ -149,45 +138,6 @@ class UBSManagementServiceImplTest {
 
     @Mock
     private BigOrderTableRepository bigOrderTableRepository;
-
-    private void getMocksBehavior() {
-
-        when(addressRepository.capacity(anyDouble(), anyDouble())).thenReturn(25);
-
-        for (Coordinates coordinate : ModelUtils.getCoordinatesSet()) {
-            List<Order> orders = ModelUtils.getOrdersToGroupThem().stream()
-                .filter(e -> e.getUbsUser().getAddress().getCoordinates().equals(coordinate)).collect(
-                    Collectors.toList());
-            when(orderRepository.undeliveredOrdersGroupThem(coordinate.getLatitude(), coordinate.getLongitude()))
-                .thenReturn(orders);
-            for (Order order : orders) {
-                when(modelMapper.map(order, OrderDto.class)).thenReturn(OrderDto.builder()
-                    .latitude(order.getUbsUser().getAddress().getCoordinates().getLatitude())
-                    .longitude(order.getUbsUser().getAddress().getCoordinates().getLongitude())
-                    .build());
-            }
-        }
-    }
-
-    @Test
-    void getClusteredCoordsTest() {
-        when(addressRepository.undeliveredOrdersCoordsWithCapacityLimit(litres))
-            .thenReturn(ModelUtils.getCoordinatesSet());
-        getMocksBehavior();
-        List<GroupedOrderDto> expected = ModelUtils.getGroupedOrders();
-        List<GroupedOrderDto> actual = ubsManagementService.getClusteredCoords(distance, litres);
-        assertEquals(expected, actual);
-    }
-
-    @Test
-    void getClusteredCoordsWithBiggerClusterLitresTest() {
-        when(addressRepository.undeliveredOrdersCoordsWithCapacityLimit(60)).thenReturn(ModelUtils.getCoordinatesSet());
-        getMocksBehavior();
-        List<GroupedOrderDto> expected = ModelUtils.getGroupedOrdersFor60LitresLimit();
-        List<GroupedOrderDto> actual = ubsManagementService.getClusteredCoords(distance, 60);
-
-        assertEquals(expected, actual);
-    }
 
     @Test
     void getAllCertificates() {
@@ -850,69 +800,6 @@ class UBSManagementServiceImplTest {
         ubsManagementService.addUserViolation(add, new MultipartFile[2], "abc");
 
         assertEquals(1, user.getViolations());
-    }
-
-    @Test
-    void getClusteredCoordsAlongWithSpecifiedTest() {
-        Coordinates coord = ModelUtils.getCoordinates();
-        Set<Coordinates> result = new HashSet<>();
-        result.add(coord);
-        List<Order> orderList = new ArrayList<>();
-        orderList.add(ModelUtils.getOrderTest());
-        when(addressRepository.undeliveredOrdersCoords()).thenReturn(result);
-        when(addressRepository.capacity(anyDouble(), anyDouble())).thenReturn(300);
-        when(orderRepository.undeliveredOrdersGroupThem(anyDouble(), anyDouble())).thenReturn(orderList);
-        when(modelMapper.map(any(), any())).thenAnswer(new Answer() {
-            private int count = 0;
-
-            public Object answer(InvocationOnMock invocation) {
-                if (count == 0) {
-                    count++;
-                    return coord;
-                }
-                return ModelUtils.getOrderDto();
-            }
-        });
-        GroupedOrderDto groupedOrderDto = ubsManagementService
-            .getClusteredCoordsAlongWithSpecified(ModelUtils.getCoordinatesDtoSet(), 3000, 15).get(0);
-        assertEquals(300, groupedOrderDto.getAmountOfLitres());
-        assertEquals(groupedOrderDto.getGroupOfOrders().get(0), getOrderDto());
-    }
-
-    @Test
-    void testAllUndeliveredOrdersWithLitersThrowException() {
-        when(addressRepository.undeliveredOrdersCoords()).thenReturn(ModelUtils.getCoordinatesSet());
-
-        List<Order> undeliveredOrders = new ArrayList<>();
-
-        when(orderRepository.undeliveredAddresses()).thenReturn(undeliveredOrders);
-
-        assertThrows(ActiveOrdersNotFoundException.class,
-            () -> ubsManagementService.getAllUndeliveredOrdersWithLiters());
-    }
-
-    @Test
-    void testGetAllUndeliveredOrdersWithLiters() {
-        List<Order> allUndeliveredOrders = ModelUtils.getOrdersToGroupThem();
-
-        when(addressRepository.undeliveredOrdersCoords()).thenReturn(ModelUtils.getCoordinatesSet());
-        when(orderRepository.undeliveredAddresses()).thenReturn(allUndeliveredOrders);
-        when(addressRepository.capacity(anyDouble(), anyDouble())).thenReturn(75, 25);
-
-        for (Coordinates cord : ModelUtils.getCoordinatesSet()) {
-            List<Order> currentOrders = allUndeliveredOrders.stream().filter(
-                o -> o.getUbsUser().getAddress().getCoordinates().equals(cord)).collect(Collectors.toList());
-            for (Order order : currentOrders) {
-                when(modelMapper.map(order, OrderDto.class)).thenReturn(
-                    OrderDto.builder().latitude(order.getUbsUser().getAddress().getCoordinates().getLatitude())
-                        .longitude(order.getUbsUser().getAddress().getCoordinates().getLongitude()).build());
-            }
-        }
-
-        List<GroupedOrderDto> expected = ubsManagementService.getAllUndeliveredOrdersWithLiters();
-        List<GroupedOrderDto> actual = ModelUtils.getGroupedOrdersWithLiters();
-
-        assertEquals(expected, actual);
     }
 
     @Test
